+++ dflib/src/test/java/org/dflib/ColumnSet_Map_RowMapperTest.java
@@ -9,12 +9,9 @@ public class ColumnSet_Map_RowMapperTest {
     public void cols_ByName() {
         DataFrame df = DataFrame.foldByRow("a", "b")
                 .of(1, "x", 2, "y")
-                .cols("b", "c").map(
-                        (f, t) -> {
-                            t.set(0, f.get(0, Integer.class) * 100);
-                            t.set(1, f.get(0, Integer.class) * 10);
-                        }
-                );
+                .cols("b", "c").map((f, t) -> t
+                        .set(0, f.get(0, Integer.class) * 100)
+                        .set(1, f.get(0, Integer.class) * 10));
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(2)
@@ -26,11 +23,9 @@ public class ColumnSet_Map_RowMapperTest {
     public void cols_ByPos() {
         DataFrame df = DataFrame.foldByRow("a", "b")
                 .of(1, "x", 2, "y")
-                .cols(1, 2).map(
-                        (f, t) -> {
-                            t.set(0, f.get(0, Integer.class) * 100);
-                            t.set(1, f.get(0, Integer.class) * 10);
-                        }
+                .cols(1, 2).map((f, t) -> t
+                        .set(0, f.get(0, Integer.class) * 100)
+                        .set(1, f.get(0, Integer.class) * 10)
                 );
 
         new DataFrameAsserts(df, "a", "b", "2")
@@ -60,12 +55,9 @@ public class ColumnSet_Map_RowMapperTest {
     public void colsExcept_ByName() {
         DataFrame df = DataFrame.foldByRow("a", "b", "c")
                 .of(1, "x", "z", 2, "y", "a")
-                .colsExcept("a").map(
-                        (f, t) -> {
-                            t.set(0, f.get(0, Integer.class) * 100);
-                            t.set(1, f.get(0, Integer.class) * 10);
-                        }
-                );
+                .colsExcept("a").map((f, t) -> t
+                        .set(0, f.get(0, Integer.class) * 100)
+                        .set(1, f.get(0, Integer.class) * 10));
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(2)
@@ -77,12 +69,9 @@ public class ColumnSet_Map_RowMapperTest {
     public void colsExcept_ByPos() {
         DataFrame df = DataFrame.foldByRow("a", "b", "c")
                 .of(1, "x", "z", 2, "y", "a")
-                .colsExcept(1).map(
-                        (f, t) -> {
-                            t.set(0, f.get(0, Integer.class) * 100);
-                            t.set(1, f.get(0, Integer.class) * 10);
-                        }
-                );
+                .colsExcept(1).map((f, t) -> t
+                        .set(0, f.get(0, Integer.class) * 100)
+                        .set(1, f.get(0, Integer.class) * 10));
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(2)