+++ dflib/src/test/java/org/dflib/DataFrame_Map_RowMapperTest.java
@@ -11,10 +11,10 @@ public class DataFrame_Map_RowMapperTest {
         Index i1 = Index.of("a", "b");
         DataFrame df = DataFrame
                 .foldByRow(i1).of(1, "x", 2, "y")
-                .map(i1, (f, t) -> {
-                    t.set(0, ((Integer) f.get(0)) * 10);
-                    t.set(1, f.get(1));
-                });
+                .compactInt(0, 0)
+                .map(i1, (f, t) -> t
+                        .set(0, f.getInt(0) * 10)
+                        .set(1, f.get(1)));
 
         new DataFrameAsserts(df, "a", "b")
                 .expectHeight(2)
@@ -27,10 +27,10 @@ public class DataFrame_Map_RowMapperTest {
         Index index = Index.of("a", "b");
         DataFrame df = DataFrame
                 .foldByRow(index).of(1, "x1", 2, "y2")
-                .map(index, (f, t) -> {
-                    t.set(0, f.get(0, Integer.class) * 10);
-                    t.set(1, f.get(1, String.class).charAt(0));
-                });
+                .compactInt(0, 0)
+                .map(index, (f, t) -> t
+                        .set(0, f.getInt(0) * 10)
+                        .set(1, f.get(1, String.class).charAt(0)));
 
         new DataFrameAsserts(df, "a", "b")
                 .expectHeight(2)
@@ -43,10 +43,10 @@ public class DataFrame_Map_RowMapperTest {
         Index index = Index.of("a", "b");
         DataFrame df = DataFrame
                 .foldByRow(index).of(1, "x1", 2, "y2")
-                .map(index, (f, t) -> {
-                    t.set(0, f.get("a", Integer.class) * 10);
-                    t.set(1, f.get("b", String.class).charAt(0));
-                });
+                .compactInt(0, 0)
+                .map(index, (f, t) -> t
+                        .set(0, f.getInt("a") * 10)
+                        .set(1, f.get("b", String.class).charAt(0)));
 
         new DataFrameAsserts(df, "a", "b")
                 .expectHeight(2)
@@ -64,12 +64,11 @@ public class DataFrame_Map_RowMapperTest {
                         Series.ofDouble(1.1, 2.2, -4.4),
                         Series.ofBool(true, false, true)
                 )
-                .map(index, (f, t) -> {
-                    t.set(0, f.getInt(0) * 10);
-                    t.set(1, f.getLong(1) * 10L);
-                    t.set(2, f.getDouble(2) * 10.);
-                    t.set(3, !f.getBool(3));
-                });
+                .map(index, (f, t) -> t
+                        .set(0, f.getInt(0) * 10)
+                        .set(1, f.getLong(1) * 10L)
+                        .set(2, f.getDouble(2) * 10.)
+                        .set(3, !f.getBool(3)));
 
         new DataFrameAsserts(df, "i", "l", "d", "b")
                 .expectHeight(3)