+++ dflib/src/test/java/org/dflib/Series_MapTest.java
@@ -17,10 +17,9 @@ public class Series_MapTest {
     @ParameterizedTest
     @EnumSource(SeriesType.class)
     public void map_DataFrame(SeriesType type) {
-        DataFrame df = type.createSeries("a", "b", "c").map(Index.of("upper", "is_c"), (v, r) -> {
-            r.set(0, v.toUpperCase());
-            r.set(1, v.equals("c"));
-        });
+        DataFrame df = type.createSeries("a", "b", "c").map(Index.of("upper", "is_c"), (v, r) -> r
+                .set(0, v.toUpperCase())
+                .set(1, v.equals("c")));
 
         new DataFrameAsserts(df, "upper", "is_c")
                 .expectHeight(3)